+++ KeyboardKit/KeyboardNavigationController.swift
@@ -21,27 +21,37 @@ open class KeyboardNavigationController: UINavigationController {
         true
     }
 
-    private lazy var leftToRightBackKeyCommands = [
-        UIKeyCommand((.command, "["), action: #selector(kbd_goBackFromKeyCommand), title: localisedString(.navigation_back)),
-        UIKeyCommand((.command, .leftArrow), action: #selector(kbd_goBackFromKeyCommand)),
-    ]
+    /// A key command that enables users to go back.
+    ///
+    /// Title: Back
+    ///
+    /// Input: ⌘`[` (Users can also use ⌘←. Both of these inputs are mirrored in right-to-left layouts.)
+    ///
+    /// UIKit has this functionality on iOS 15, but KeyboardKit does this back to iOS 12.
+    /// This API is only available from iOS 15 because using a single key command for this requires automatic
+    /// mirroring, which was not available until iOS 15. However KeyboardKit provides this functionality back
+    /// to iOS 12 using multiple internal key commands to implement mirroring.
+    @available(iOS 15.0, *)
+    public static let backKeyCommand = discoverableLeftToRightBackCommand
 
-    private lazy var rightToLeftBackKeyCommands = [
-        // Note that on iOS 14 and earlier the system will incorrectly show this in the discoverability HUD as a leftwards
-        // pointing arrow. The discoverability HUD mirrors the arrow keys it displays when in a right-to-left layout, but
-        // the inputs on the actual events received are not mirrored. This was reported as FB8963593 and resolved in iOS 15.
-        UIKeyCommand((.command, "]"), action: #selector(kbd_goBackFromKeyCommand), title: localisedString(.navigation_back)),
-        UIKeyCommand((.command, .rightArrow), action: #selector(kbd_goBackFromKeyCommand)),
-    ]
+    private static let discoverableLeftToRightBackCommand = DiscoverableKeyCommand((.command, "["), action: #selector(kbd_handleCmdLeftBracket), title: localisedString(.navigation_back))
+    // Note that on iOS 14 and earlier the system will incorrectly show this in the discoverability HUD as a leftwards
+    // pointing arrow. The discoverability HUD mirrors the arrow keys it displays when in a right-to-left layout, but
+    // the inputs on the actual events received are not mirrored. This was reported as FB8963593 and resolved in iOS 15.
+    // TODO: Update this comment since the discoverable one was switched to ] so maybe that doesn’t have the same problem.
+    private static let discoverableRightToLeftBackCommand = UIKeyCommand((.command, "]"), action: #selector(kbd_handleCmdRightBracket), title: localisedString(.navigation_back))
 
-    private var backKeyCommands: [UIKeyCommand] {
+    private static let nonDiscoverableLeftToRightBackCommand = UIKeyCommand((.command, .leftArrow), action: #selector(kbd_goBackFromKeyCommand))
+    private static let nonDiscoverableRightToLeftBackCommand = UIKeyCommand((.command, .rightArrow), action: #selector(kbd_goBackFromKeyCommand))
+
+    private var shouldUseRightToLeftBackCommand: Bool {
         if #available(iOS 15.0, *) {
             // Handled by allowsAutomaticMirroring.
-            return leftToRightBackKeyCommands
+            return false
         } else {
             switch view.effectiveUserInterfaceLayoutDirection {
-            case .rightToLeft: return rightToLeftBackKeyCommands
-            case .leftToRight: fallthrough @unknown default: return leftToRightBackKeyCommands
+            case .rightToLeft: return true
+            case .leftToRight: fallthrough @unknown default: return false
             }
         }
     }
@@ -63,9 +73,18 @@ open class KeyboardNavigationController: UINavigationController {
                 (systemCommand.input == "[" && systemCommand.modifierFlags == .command) == false
             }
 
-            // It’s useful to check this otherwise these commands can block rewind/fast-forward bar button items unnecessarily.
+            // It’s useful to check canGoBack otherwise these commands can block rewind/fast-forward bar button items unnecessarily.
+            // Although when using the menu builder we can’t do anything about that.
             if canGoBack {
-                additionalCommands += backKeyCommands
+                if shouldUseRightToLeftBackCommand {
+                    additionalCommands.append(Self.nonDiscoverableRightToLeftBackCommand)
+                    additionalCommands.append(Self.discoverableRightToLeftBackCommand)
+                } else {
+                    additionalCommands.append(Self.nonDiscoverableLeftToRightBackCommand)
+                    if Self.discoverableLeftToRightBackCommand.shouldBeIncludedInResponderChainKeyCommands {
+                        additionalCommands.append(Self.discoverableLeftToRightBackCommand)
+                    }
+                }
             }
 
             let keyCommandFromBarButtonItem: (UIBarButtonItem) -> UIKeyCommand? = {
@@ -89,20 +108,31 @@ open class KeyboardNavigationController: UINavigationController {
     }
 
     open override func canPerformAction(_ action: Selector, withSender sender: Any?) -> Bool {
-        if action == #selector(kbd_goBackFromKeyCommand) {
-            return canGoBack
-        }
+        // The menu builder provides both the LtR and RtL commands, so block the wrong one. For this reason, the menu
+        // builder support here can’t be used on Catalyst since it would show the available command as disabled.
 
-        return super.canPerformAction(action, withSender: sender)
+        switch action {
+        case #selector(kbd_goBackFromKeyCommand):
+            return canGoBack
+        case #selector(kbd_handleCmdLeftBracket):
+            return shouldUseRightToLeftBackCommand ? false : canGoBack
+        case #selector(kbd_handleCmdRightBracket):
+            return shouldUseRightToLeftBackCommand ? canGoBack : false
+        default:
+            return super.canPerformAction(action, withSender: sender)
+        }
     }
 
     private var canGoBack: Bool {
         viewControllers.count > 1 && presentedViewController == nil && navigationItem.hidesBackButton == false && (navigationItem.nnLeadingBarButtonItems.isEmpty || navigationItem.leftItemsSupplementBackButton)
     }
 
+    @objc private func kbd_handleCmdLeftBracket(sender: UIKeyCommand) { self.kbd_goBackFromKeyCommand(sender) }
+    @objc private func kbd_handleCmdRightBracket(sender: UIKeyCommand) { self.kbd_goBackFromKeyCommand(sender) }
+
     // Important to not put this on all UINavigationController instances via an extension because those
     // instances lack our override of canPerformAction so could allow the action incorrectly.
-    @objc func kbd_goBackFromKeyCommand(_ keyCommand: UIKeyCommand) {
+    @objc private func kbd_goBackFromKeyCommand(_ keyCommand: UIKeyCommand) {
         let allowsPop = navigationBar.delegate?.navigationBar?(navigationBar, shouldPop: topViewController!.navigationItem) ?? true
         guard allowsPop else {
             return